UPDATE AND DELETE USER CRUD

+++ Tienda_proyecto/src/main/java/Tienda_proyecto/InterfaceService/IUserService.java
@@ -7,8 +7,8 @@ import Tienda_proyecto.Models.tbusuarios;
 
 public interface IUserService {
 	public List<tbusuarios> List();
-	public Optional<tbusuarios> Find(int id);
+	public Optional<tbusuarios> Find(long id);
 	public int Create(tbusuarios data);
-	public int Delete(int Id, int Mod);
+	public int Delete(long Id, long Mod);
 	public int Update(tbusuarios data);
 }
